chore(inboxes): fixed wobbly tests

+++ packages/inboxes/migrations/20170926113900_create_inbox_conversation_table.js
@@ -1,27 +1,28 @@
 'use strict';
 
-exports.up = knex => {
+exports.up = async knex => {
   const { schemas } = knex.client.config;
 
-  return knex.schema.createTableIfNotExists(
-    schemas.inboxConversation,
-    table => {
-      table.charset('utf8');
-      table.collate('utf8_general_ci');
+  if (await knex.schema.hasTable(schemas.inboxConversation)) {
+    return;
+  }
 
-      table.bigInteger('inbox_id').unsigned().notNullable().index();
-      table.bigInteger('conversation_id').unsigned().notNullable().index();
+  return knex.schema.createTable(schemas.inboxConversation, table => {
+    table.charset('utf8');
+    table.collate('utf8_general_ci');
 
-      table
-        .foreign('conversation_id')
-        .references(`${schemas.conversation}.id`)
-        .onDelete('CASCADE');
-      table
-        .foreign('inbox_id')
-        .references(`${schemas.inbox}.id`)
-        .onDelete('CASCADE');
-    }
-  );
+    table.bigInteger('inbox_id').unsigned().notNullable().index();
+    table.bigInteger('conversation_id').unsigned().notNullable().index();
+
+    table
+      .foreign('conversation_id')
+      .references(`${schemas.conversation}.id`)
+      .onDelete('CASCADE');
+    table
+      .foreign('inbox_id')
+      .references(`${schemas.inbox}.id`)
+      .onDelete('CASCADE');
+  });
 };
 
 exports.down = knex => {
